Abstracted common code from hierarchical phrases into abstract class.
Preparing to remove HierarchicalPhrase class. This is being done for efficiency (and hopefully some speedup). There are a lot of HierarchicalPhrase objects that are created - it should take relatively little effort to merge the functionality of this class into MatchedHierarchicalPhrases. This will be especially useful as MemoryMappedHierarchicalPhrases is developed for very frequent phrases, as this new approach will alleviate the requirement to create actual HierarchicalPhrase objects, with data copied from the memory-mapped data structure.

Also minor work on calculating the very frequent phrases. SuffixArrayWithExtras needs a better name.

git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@643 0ae5e6b2-d358-4f09-a895-f82f13dd62a4