More work on integrating suffix array code into decoder.
The main method in the new Compile class can now be used to extract all necessary binary files from an aligned parallel corpus. These files will be placed into a common directory which should end in a .josh extension. The actual binary files will have specific, meaningful names within the directory. This will mean that when using suffix array grammars (extracted on-the-fly), the user will be able to simply specify the .josh directory as the tm_grammar file.


git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@673 0ae5e6b2-d358-4f09-a895-f82f13dd62a4