Changes to enable greater compatibility with Hiero's rule extraction.
In ExtractRules, the following changes have been made:

* --maxPhraseSpan now defaults to 5 (previously defaulted to 10, which was just plain a bad default value)

* --sentence-initial-X now defaults to true (to match what Hiero does)
* --sentence-final-X now defaults to true (to match what Hiero does)

* new option --violating-X added, which defaults to true (to match what Hiero does)

This final option allows rules with an initial (or a final) X to be extracted, even in cases where the total span of the phrase is longer than maxPhraseSpan. Hiero's grammar extraction code appears to behave this way; it seems to only consider the maximal terminals of a phrase when determining if a phrase violates the maxPhraseSpan. The code to implement this in Joshua may not be perfect, but it appears to work in the cases tested so far.

Unit tests for ExtractRules have been expanded to test these new cases.

Documentation for the above Hiero-compatibility options have been added to the documentation.



git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@707 0ae5e6b2-d358-4f09-a895-f82f13dd62a4