Lexprobs bug fix.
The bug was in the HierarchicalPhrase constructor, caused by failure to increment a loop counter. This is why Java needs Ruby-style iteration loops, that allow each_with_index style iteration!

int terminalSpanNumber = 0;
for (Span terminalSpan : terminalSpans) {
	terminalSequenceStartIndices[terminalSpanNumber] = terminalSpan.start;
	terminalSequenceEndIndices[terminalSpanNumber] = terminalSpan.end;
	terminalSpanNumber++;
}

This should be writable as something like:

for (Span terminalSpan, int terminalSpanNumber : terminalSpans) {
	terminalSequenceStartIndices[terminalSpanNumber] = terminalSpan.start;
	terminalSequenceEndIndices[terminalSpanNumber] = terminalSpan.end;
}

I have not yet verified that this bug fix resolves all outstanding differences between lexprobs for hiero-extracted grammars versus Joshua-extracted grammars, but it definitely fixes the case for "que X . ||| that X ." when training on the first five sentences of Europarl es-en and extracting for the first sentence in dev2006.

git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@893 0ae5e6b2-d358-4f09-a895-f82f13dd62a4