(This is a long commit note, since this is change is significant in a certain way.)
Reorganized the code that processes the decoder output files.  SS are
now calculated only after determining new candidates for *all* the
source sentences, instead of querying the SS method once per source
sentence.  To keep memory requirement low, there is some reliance on
disk and some extra I/O.  (The extra overhead is quite low, on the
order of seconds per iteration.)

Why do this?  Because if the SS method relies on an external process,
as is the case with tercom for TER, the old structure spawns an
external process per source sentence.  So there is some overhead for
launching too many processes, now being eliminating.  Also, if the
external evaluator starts by loading/creating structures such as
a language model or a synonym table, we are eliminating the need to
do this once per source sentence.  The TER class has been modified
accordingly.

As a side effect, the new structure also makes parallelizing TER
computation easier, since TER.java now queries tercom with batches of
candidates (10k candidates per batch; should probably be made a TER
option).  The batches are created first, and only then is tercom
launched, processing each batch file sequentially, and so this should
be relatively easy to parallelize.


git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@1127 0ae5e6b2-d358-4f09-a895-f82f13dd62a4