Fixed bug introduced in r1291.
The changes in r1291 were intended to better handle (by properly ignoring) sentences which were too long for grammar extraction to handle. This is fine, except that the pointCounter value for ignored long sentences was incorrectly writeExternal'ed as zero in r1291. The pointCounter variable is defined as a strictly non-decreasing counter value. Thus, the proper way to ignore a sentence, as far as pointCounter is concerned, is to increment by zero (that is, do nothing) rather than resetting the value to zero.

Unfortunately, it appears that proper unit tests have not been written to check for this case. This needs to be done.

git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@1710 0ae5e6b2-d358-4f09-a895-f82f13dd62a4