"thrax" grammar format keyword for joshua
Since thrax format is just hiero format, internally it calls the hiero
reader. But this should make it clearer for people who are using thrax
and joshua in combination. Also changed the joshua.config template to
use the thrax keyword. It was originally a <GRAMMAR> variable that
gets replaced, but that breaks if using thrax to extract an SAMT grammar.
Joshua would want samt format in that case, which thrax does not do.