fixed lattice scoring
The SourcePathFF class had not implemented reEstimateTransitionLogP, which is the one
that is called by the cube prune state combiner.  I chained this to transitionLogP, so
the lattice score is now incorporated as part of the model score.

I had to turn off sanity checking in the kbest extractor.  I get errors when the
lattice-cost feature is enabled, but I do not follow the code as written here.  This whole
class needs to be rewritten so that it is understandable (k-best hypergraph extraction).