removed OOV handling from the grammars, and simplified the Rule interface
Before, OOV handling was done by the grammars, which knew how to create OOV
rules.  This didn't really make sense (e.g., the packed grammar can't add rules
on the fly), so now that we can handle multiple grammars, we simply create a new
grammar in the chart initialization code to hold that sentence's OOV rules.

The Rule interface was also simplified, removed the unusued lattice and rule
ID variables.