updated to work with sparse feature version, but bug remains in global Vocabulary
The sparse feature representation should work now with the Packed Grammar, except for the
fact that the PackedGrammar class resets the vocabulary, invalidating calls made before
this grammar is instantiated, and limiting the number of packed grammars to one.  This
wrecks the score computation because the owners don't match on the grammars.

To fix this, each packed grammar needs its own private vocabulary and should map
between that and the global vocabulary.