Updated test cases with sparse grammar implementation.
The tests matched exactly on the bn-en/samt test.  There were some discrepancies on about
10% of the sentences for the bn-en/hiero test, but they went both ways.  For every sentence
that the old version had a higher model score, constrained decoding under the new implementation
produced the new score.  In light of this, the success of the SAMT test, and Kenneth Heafield's
tests that showed the sparse implementation getting slightly higher scores for a given pop
level than devel, I'm willing to accept the new scores as gold, and consider the differences
to be due to random pruning effects.

I also updated the test cases to compare exact output and model scores across the whole
n-best list, instead of just computing a BLEU score.  This is what we should have been
doing all along.