Rewrote the internals of thread management.
Before, we used an InputHandler which knew about all of the file input.  The threads
were created with knowledge of the InputHandler, and queried it for sentences until
it was exhausted, at which point they died.

This does not work for a model where there may be multiple (many-sentence) requests,
such as for a server version.  So we've rewritten things to use a thread pool.  Requests
(chunks of sentences) now continually obtain and release threads until all of their
sentences are translated.