synchronous parsing: use Strings for node labels
Matt claims that with forced decoding on the first pass, the
chart will be small enough that we don't need to do all this
hacky stuff when creating the second-pass grammar.

So instead we just do the simple thing, and add strings of the
form "i-X-j" to the Vocabulary and use them for the grammar.
Originally, we didn't explicitly build up the strings, which led
to ugly bookkeeping.