Bugfix in the server threading
Previously, the server would continue to translate all the sentences of an aborted
TCP/IP connection, even though it could no longer return them.  This occurred because
the input sentences appear to have been queued prior to the client termination (likely
because of the buffered input), so the TranslationRequest could continue to get sentences
to translate and then translate them.

The fix was to trap socket exceptions and tell the TranslationRequest object to stop
returning sentences.