Old code reversed the sign of weights in the lattice, which I think is a bad idea.
Note that the phrase table code does the same thing: values found in the grammars
are all multiplied by -1.  At some point in the future it would be nice to get rid
of this since it just introduces a hidden piece of indirection which complicates
evaluating model scores.  Changing that would result in loss of backwards compatibility.
Here I am not worried about that because the lattice code is relatively new and
no one is using it.