Removed call to Vocabulary::word() in computation of LM portion of cell signature.
This fixes #92.

The problem was twofold: (a) research code computing a nice-looking string signature
including words, unnecessarily (perhaps conflating the hash code and the string rep
of the cell), and (b) calls to Vocabulary::word() are synchronized (whcih probably
doesn't need to be the case).

We recommend a full review of hashing code to get rid of research-motivated shortcuts like
this, as well as strict attention paid to hashCode(), str(), and equals() (each should have
their proper role).