(1) Folded state computers into feature functions and (2) generalized feature function interface
Regarding (2), features no longer define both

  float computeCost()
  FeatureFunction computeFeatures()

Instead, they define only compute(), which takes a new accumulator object that accumulates
values in the style required (save just the cost or the result) and returns them to the
caller in the desired fashion.

Regarding (1), we no longer have separate state computers. Stateful features now know how to
compute their own state. State can't be shared across, say, LMs this way, but it saves us
making two passes across the rules, one for state computation, and one for score computation.
compute() for a feature function returns the state, and the accumulated value is stored as
an object passed to it for retrieval afterwards.