Deprecated old feature-file loading format and updated test cases.
In the past, the presence of a feature weight would trigger a feature (e.g., OOVPenalty), and many
features could be specified in one way in the config file and another way in the weights file
(e.g., "phrasemodel pt 0 WEIGHT" and "tm_pt_0 WEIGHT". Now, features have to be triggered only
by the presence of the "tm" or "lm" lines or an explicit invocation such as "feature-function = OOVPenalty".
Weights can be listed in either the config file or the weights file. The weights file was introduced
last year but it's more trouble than it's worth, so I'll be pushing to just have the entire model
in a single file.