Added -rescore-forest option, removed -oracle-file
If you want to pass in an oracle / reference sentence, it is the third |||-delimited
field on STDIN to Joshua. Probably some effort should go into rethinking the STDIN
input format; currently the second field is used for a target-language parsing or
forced-decoding target, which is why we used the third field here.