Gideon : Fixed  test/joshua/decoder/FeatureFunctionsTest.java so the test works out of the box, without manually setting the right Djava.library.path in the VM arguments of eclipse. This is done by a special function that extends "usr_paths" using reflection.
Also adapted the .classpath. Added the junit jar, which is now also retrieved from
the ivy dependencies. Removed includes of source / javadoc jars from mockito
and testng that are not automatically downloaded, and then have to every time
be manually deleted in eclipse, which is annoying.
Remark: Another way to fix this is to change the ivy.xml somehow to fetch the
javadoc / sources automatically as well, but not sure how to do this.