Gideon : Added a new feature function LabelSubstitutionFF.java which is to be used in combination with soft-syntactic decoding. The feature function fires boolean features for:
1. The match/mismatch of the rule label with the substituted label
2. A particular subsitution of a substituted label for a certain rule label

Adapted the tests FeatureFunctionTest to test if this feature is firing and
SoftSyntacticConstraintDecodingTest to test if this feature yields different
feature activations if soft syntactic constraint decoding is used as compared
to when normal decoding is used.

Added new ckass src/joshua/decoder/ff/RulePropertiesQuerying.java
to put shared static method used by different feature funtions in a central place.

       modified:   src/joshua/decoder/Decoder.java
       modified:   src/joshua/decoder/ff/LabelCombinationFF.java
       new file:   src/joshua/decoder/ff/LabelSubstitutionFF.java
       new file:   src/joshua/decoder/ff/RulePropertiesQuerying.java
       modified:   test/joshua/decoder/FeatureFunctionsTest.java
       modified:   test/joshua/decoder/SoftSyntacticConstraintDecodingTest.java