Fixed the names of the feature function that are added: they must be consistently lowercased for consistency with the FeatureVector class and its weights table.
Lowercasing the feature fucntion names in the method  initializeFeatureFunctions()
in the Decoder.java class, but then still using uppercased feature names in the feature classes themselves
lead to inconsistency and nasty bugs were setting the weight of certain features like OOV and WordPenalty
had no effect. By also using the lowercased version in the constructors of these classes, this bug
is now fixed.

   modified:   src/joshua/decoder/ff/ArityPhrasePenaltyFF.java
   modified:   src/joshua/decoder/ff/OOVFF.java
   modified:   src/joshua/decoder/ff/SourcePathFF.java
   modified:   src/joshua/decoder/ff/WordPenaltyFF.java