Pipeline now handles mixed dense and sparse features in any grammar
This required extending the EncoderConfiguration to print out feature names,
and then generalizing the pipeline to use those feature names in the initial
config files that it generates for tuning. Default values for everything are 1;
it might be worthwhile allowing sparse features to be initialized to 0, either
hard-coded or as an option.