Straightened out annoying polarity issues with score representation
Joshua has multiplications by -1 all over the place, making it really
hard to follow what's going on. Some of this is a holdover to when
the model was more strictly probabilistic. Now, we always represent
scores as (weight vector * feature values), and report them directly
(whereas before sometimes they were negated and reversed on storage
and printout).

IMPORTANT: Joshua still negates the values of features it reads in,
for backwards compatibility (Thrax outputs features that have been
mapped through -log(x)). However, Joshua *no longer* negates the
values of sparse features found in the grammar file. It used to do
this.