BUGFIX: invalid use of "i", fixed Hypothesis equals()
Was using HGNode's "i" parameter to keep track of the beginning of a phrase, but this isn't valid, because i is not part of the DP state. Corrected this by computing i where needed by subtracting off the number of words on the rule's source side.

Also (important) fixed the Hypothesis equals() code to compare DP states