Update test cases to work after fix from commit 7565550748ef771ea34c1745556f73dc921206d5
That commit fixed the distance() function for sentences, which up till then was returning Integer.MIN_VALUE, causing there to be effectively no limit on the length of spans to which rules were applied. Interestingly, decoding with no limit on the span doesn't affect decoding speed all that much.

Although there are many changes to the gold files here, I trust them for the following reasons:

- A good many of them are very small floating point rounding errors
- I decoded a large WMT13 test set and found
    - (a) improved model score over the current master commit (commit a4b4d150bdfb814c4264933296f88f52773f774a)
    - (b) BLEU score