KenLM now computes rule estimates itself
Previously, multiple calls were made to KenLM::probString(), for each ngram in the rule. This is the generic LM interface. Now, a single call is made to KenLM::estimateRule(). For terminal rules, this avoids some setup in the call chain, and for rules with nonterminals, it potentially avoids many calls over the JNI interface.

The real benefits here will therefore accrue to Hiero and GHKM decoding.