Fixed alignment computation
Alignments are now stored as strings for unpacked grammars, and transformed into byte[] arrays only on demand. This commit includes a new rule type for Moses phrase table rules, so that the alignment offset of 1 (necessary because of the prepended nonterminal) can be computed on the fly without too much mess.