Merged LatticeInput and Sentence, now translating Tokens
General cleanup of lattice and sentence interfaces. The decoders now always translate lattices. A Sentence should be a special case of a lattice.

The lattice is now composed of Tokens, not integers. A token can have an annotation and maybe some other stuff.

Language model not yet using Token classes.