Removed slow and redundant feature string parsing when constructing rules from packed grammar (at sort time and at actual construction of feature vector).
Gets rid of String parsing features over and over again which turned out to be slow in profiling. The solution is not perfect, but we get a nice speedup of roughly a factor 5: If JoshuaConfiguration.amortize is set to false grammars are forced to be sorted at decoder startup. Here are the stats: New code: Took 561.64 seconds to load pipeline. Old code: Took 2688.60 seconds to load pipeline.  Basically we are significantly reducing the time for sorting the rules by getting rid of an intermediate string representation of the features in a rule. Since String parsing of floats is removed now there was some float precision change in the regression-test for which I changed the gold output. This is fine.