updating handling of source-side annotations
Joshua now permits source-side annotations. Each word can have attached an unlimited number of ;-delimited key[=value] annotations attached to each input word, in brackets. For example:

  je[pos=PRO] suis[pos=VB,tense=present;irregular] ...

This will result in annotations being attached to each word, which are then available to feature functions and elsewhere in the decoder. If an annotation has no value, it because its own value.

If -source-annotations is passed to the decoder, annotations of the name "class" will be used in place of the target-side word, allowing a language model to key on source-side context.